+++ ferrodec-decimal64/src/lib.rs
@@ -25,6 +25,7 @@
 mod bid;
 mod classify;
 mod classify_types;
+mod cmp;
 mod convert;
 mod decimal;
 mod ops;